% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sqrt_invsqrt}
\alias{sqrt_invsqrt}
\title{Function used to obtain the square root and inverse square root of an input matrix.}
\usage{
sqrt_invsqrt(matrix, symmetric, printInfo = FALSE)
}
\arguments{
\item{matrix}{- the input matrix to be used in eigenvalue decomposition. It is assumed to be square}

\item{symmetric}{- forces the user to make a choice about the symmetry of the input matrix.}

\item{printInfo}{- Prints diagnostic information about the client processing}
}
\value{
A list that contains the sqrt and the inverse sqrt of the input matrix
}
\description{
This function returns the square root and inverse square root of an input matrix.
The method involves the offload of the matrix data to a seperate syevd_server executable by copying data into a shared memory area and
signalling to the server that the data is availble. This function will block the R side code until the server has completed the decomposition. The
function checks that the input is square, however it *does not* check that the matrix is symmetric.
N.B. The maximum size allowed of the input matrix is goverend by what was provided in the rcppMagmaSYEVD::RunServer() function.
The server will need to be restarted with a larger shared memory area if user wants to perorm EVD on a larger matrix.
}

